changed chunk size constant from 1 to 1L to avoid cast to int in long constant, allowing larger chunk size up to Integer.MAX_VALUE.
git-svn-id: https://svn.code.sf.net/p/jaer/code/jAER/trunk@9337 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/net/sf/jaer/eventio/AEFileInputStream.java
@@ -156,10 +156,10 @@ public class AEFileInputStream extends DataInputStream implements AEFileInputStr
      * the size of the memory mapped part of the input file. This window is
      * centered over the file position except at the start and end of the file.
      */
-    private int CHUNK_SIZE_EVENTS = 1 << 25;
-    private int chunkSizeBytes = CHUNK_SIZE_EVENTS * EVENT32_SIZE; // size of memory mapped file chunk, depends on event
+    private long CHUNK_SIZE_EVENTS = 1L << 25;
+    private long chunkSizeBytes = CHUNK_SIZE_EVENTS * EVENT32_SIZE; // size of memory mapped file chunk, depends on event 
     // size and number of events to map, initialized as
-    // though we didn't have a file header
+    // though we didn't have a file header. Max value is Integer.MAX_VALUE however. Will generate illegalargument exception if we try to allowcate larger chunk
 
     /**
      * the packet used for reading events.
